% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{assessCOA}
\alias{assessCOA}
\title{Assess a COA design for the usual COA properties}
\usage{
assessCOA(coades, ...)
}
\arguments{
\item{coades}{The design to be assessed.}

\item{...}{
  Arguments passed on to \code{\link[=ckCOA]{ckCOA}}
  \describe{
    \item{\code{silent}}{If \code{TRUE}, a message is printed only if the check fails,
not if it succeeds.}
  }}
}
\value{
\code{TRUE} if the design passes all checks, \code{FALSE} if not.
}
\description{
This function is a friendly wrapper for \code{\link[=ckCOA]{ckCOA()}}.  It accepts a
purported COA in
either 0-based nrow>ncol format or 1-based ncol>nrow format.
}
\seealso{
\code{ckCOA} for details of the checks that are made.
}
