% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{assessOA}
\alias{assessOA}
\title{Assess an orthogonal array for validity at strength 2}
\usage{
assessOA(oades, ...)
}
\arguments{
\item{oades}{The orthogonal array to be assessed.}

\item{...}{
  Arguments passed on to \code{\link[=ckOA]{ckOA}}
  \describe{
    \item{\code{silent}}{If \code{TRUE}, a message is printed only if the check fails,
not if it succeeds.}
  }}
}
\value{
\code{TRUE} if the OA is valid, \code{FALSE} if not.
}
\description{
This function is a friendly wrapper for \code{\link[=ckOA]{ckOA()}}.  It accepts an OA in
either 0-based nrow>ncol format or 1-based ncol>nrow format.
}
\seealso{
\code{ckOA} for details of the checks that are made.
}
