% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils1.R
\name{checkWeights}
\alias{checkWeights}
\title{Check whether respondent weights are OK}
\usage{
checkWeights(weights, nresp, caller = "??")
}
\arguments{
\item{weights}{The weights vector to be checked.}

\item{nresp}{Number of respondents and expected number of weights.}

\item{caller}{Name of function calling us, for use in messages/errors.}
}
\value{
\code{TRUE} if weights look OK, \code{FALSE} if any problem found.
}
\description{
This function creates an error if weights are not of right length
or any are negative.
}
\keyword{internal}
