% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{ckCOA}
\alias{ckCOA}
\title{Check a Component Order of Addition design for key properties}
\usage{
ckCOA(coades, silent = FALSE)
}
\arguments{
\item{coades}{The COA array to be checked, in 0-coded nrow>ncol format}

\item{silent}{If \code{TRUE}, a message is printed only if the check fails,
not if it succeeds.}
}
\value{
\code{TRUE} if the COA passes all checks, \code{FALSE} if not.
}
\description{
This function checks a COA design for equi-positioning and
pairwise equi-precedence.
}
\details{
This is not meant to be called directly by a user (see \code{\link[=assessCOA]{assessCOA()}}
instead).

Various diagnostic messages are sent to the console in the event of failure.

COAs that have been size-limited (the orderings all shortened) or formed by
dropping more than one item from a "true" COA will not pass all checks,
but are still acceptable for most use cases.

It is possible a design could pass
both tests without being an actual COA.  In particular, "Williams"
designs will pass.  However, they do not resemble COAs at all, and
this function is only intended as a safety check on COA generation,
not to catch all possible frauds!
}
\keyword{internal}
