% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils1.R
\name{ckMinMax}
\alias{ckMinMax}
\title{Check validity of, and default, \code{minitems} and \code{maxitems} parameters to
drivers}
\usage{
ckMinMax(
  minitems = NULL,
  maxitems = NULL,
  orders = NULL,
  nitems = NULL,
  caller = "??"
)
}
\arguments{
\item{minitems}{Minimum combination size to use for
size-limited Shapley Values.}

\item{maxitems}{Minimum combination size to use for
size-limited Shapley Values.}

\item{orders}{An orderings matrix, data matrix or SCP matrix from which
the number of items can be deduced.}

\item{nitems}{The number of items or variables in the problem or dataset;
must be specified by name if \code{orders} is omitted.}

\item{caller}{Name of function calling us, for use in messages/errors.}
}
\value{
A 2-item numeric vector with the final values of
minitems and maxitems.
}
\description{
This function checks the legitimacy of \code{minitems} and \code{maxitems} parameters
provided to SV-calculation drivers.  It also defaults them as
appropriate and returns their final values.
The intention is to make handling of these parameters consistent
across drivers, while avoiding duplication of code.
Invalid values produce a "stop" with a message.

Ideally, \code{nitems} will be provided by the caller.
}
\details{
An alternative is to provide a data matrix, SCP matrix or orderings (COA)
matrix from which the total number of items can be deduced.
Specifically:
Usually either a raw data array with as many columns as items
(assumed if there are more rows than columns),
or a sum-of-crossproducts matrix with as many rows and columns as items
plus one (assumed if there are as many rows as columns;
note that this will not work if there are multiple dependent variables
in a key drivers regression),
or an orderings array (such as a COA) with either as many rows as items
or as many rows as the maxitems value desired
(assumed if it has more columns than rows).
}
\keyword{internal}
