% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{ckOA}
\alias{ckOA}
\title{Check an orthogonal array to be sure it is OK (strength 2 only)}
\usage{
ckOA(oades, silent = FALSE)
}
\arguments{
\item{oades}{The orthogonal array to be checked, in 0-coded nrow>ncol format.}

\item{silent}{If \code{TRUE}, a message is printed only if the check fails,
not if it succeeds.}
}
\value{
\code{TRUE} if the OA is valid, \code{FALSE} if not.
}
\description{
This function checks for equal pairwise precedence counts and for
uncorrelated columns.
}
\details{
This is not meant to be called directly by a user (see \code{\link[=assessOA]{assessOA()}} instead).
}
\keyword{internal}
