% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{cpBuild}
\alias{cpBuild}
\title{Construct a cross-product matrix for use in various key driver routines}
\usage{
cpBuild(
  df,
  depvar = NULL,
  precenter = TRUE,
  weights = NULL,
  silent = FALSE,
  missing = c("none", "listwise", "pairwise")
)
}
\arguments{
\item{df}{A matrix or data frame of the X's and Y's for the result matrix.}

\item{depvar}{Numeric vector of dependent variable columns in \code{df} OR
character vector of dependent variable names.  If \code{NULL}, the
final column of \code{df} is assumed to be the only dependent variable.}

\item{precenter}{Logical, whether to pre-center all variables and omit any
constant row/column from the result.}

\item{weights}{Vector of respondent weights, one per row of \code{df},
or character name of the weights column in \code{df}.
Unweighted data if \code{NULL}.
Weights may be zero to subset data but may never be negative.}

\item{silent}{Logical, whether to avoid sending commentary to console as
the CP matrix is constructed.}

\item{missing}{If "none", any missing data in \code{df} is an error.
If "listwise", any row/observation/respondent with any missing data
is dropped from the CP matrix.
If "pairwise", missing data is excluded only from the specific pairs
of variables that include missings.  This can sometimes result in a
cross-products or covariance matrix that is not positive definite
and is thus not valid for regression,
as diagnosable by a negative eigenvalue for the SCP matrix,
in which case an error is thrown.}
}
\value{
A square, symmetric cross-products matrix.  If \code{precenter} is
\code{FALSE}, the first row/column is for a constant variable, whether or not
one was found in \code{df}.  The dependent variables are always the last
row/columns.  Returns \code{NULL} if \code{weights} or \code{depvar} are invalid.
}
\description{
This function builds the cross-product matrix needed for key driver
regressions.
It sets up for multiple dependent variables (always the
last rows/columns of the result) and can either include a constant
row/column (always as the first row/column)
or pre-center all variables to avoid the need for one.

If \code{df} contains any non-numeric columns, they are ignored.  If there
are constant columns, only the first is used.
}
\details{
The matrix produced has an \code{ESS} attribute giving Kish's effective sample
size (square of sum of weights, over sum of squares of weights) that can
be used as the total degrees of freedom when computing adjusted r-squared.
}
