% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{cpConstantOut}
\alias{cpConstantOut}
\title{"Sweep out" and drop the constant row/column of a cross-products matrix}
\usage{
cpConstantOut(cp, constant = 1)
}
\arguments{
\item{cp}{A cross-products matrix (square, symmetric).}

\item{constant}{The row/column number of the constant, often 1.}
}
\value{
A matrix with one less row and column than \code{cp}, with the
Gauss-Jordan sweep (or Beaton's \code{swp}) via \code{\link[=swp]{swp()}} for the constant
having been done.
The result is still square, but no longer symmetric.
}
\description{
This function removes the constant from a cross-products matrix, preparing
it for use in various key driver regression situations.  Since the
constant does not affect r-squared explained, its presence is simply a
nuisance when computing r-squareds and increments.
}
\details{
An easy alternative to calling this function is to use \code{precenter=TRUE}
with \code{\link[=cpBuild]{cpBuild()}}, which eliminates any constant from the matrix.

If matrix \code{cp} has an \code{ESS} attribute giving the effective sample size
(see \code{\link[=cpBuild]{cpBuild()}}), it is preserved in the result.
}
