% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{designCode}
\alias{designCode}
\title{Convert orthogonal array or COA design to internal "0-based nrow>ncol" format}
\usage{
designCode(des)
}
\arguments{
\item{des}{A design or orthogonal array (numeric matrix) to be converted.}
}
\value{
The design with 0-based coding, with any constant first column
of ones removed, and with runs or orders in the rows, and items in the
columns (nrow>ncol format).
}
\description{
This converts an orderings or combinations design
\itemize{
\item to use 0-based numbering of items,
\item to eliminate any constant 0 or 1 entry in all first positions, and
\item to have runs/orders/combinations in the rows and items in the columns
(meaning the number of rows almost always exceed the number of columns).
}

The design is examined to see which of the output criteria is does not
already meet.  An already-compliant design is returned unchanged.
}
\details{
Aside from development and testing, it is not clear what the use case for
this would be!

The possibility of a leading 0 or 1 arises from designs used for logistic
regression, where the constant must always be entered first and
explicitly.
}
\seealso{
\code{\link[=userCode]{userCode()}} to do the reverse.
}
