% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{dropCOA}
\alias{dropCOA}
\title{Drop excess items from a COA}
\usage{
dropCOA(coades, nvars)
}
\arguments{
\item{coades}{The COA from which items are to be dropped, in
0-coded nrow>ncol format.}

\item{nvars}{The number of items/variables to keep.}
}
\value{
A new, reduced, Component Order of Addition (COA) design,
in the 0-based nrow>ncol format, with items numbered nvars or higher
(0-based) removed.  This result is not a "true" COA; it will have
pairwise equi-precedence, but will not have equi-positioning if more
than one item was removed.  The result has the same number of rows as
the original, but only \code{nvars} columns.
}
\description{
This function removes extra items from a COA when the generated COA
(typically from \code{makeCOA}) has more items than were desired.
}
\details{
This is not meant for direct calls by the user; it uses the 0-based
nrow>ncol format and will not work on COAs delivered to the user by
\code{getCOA}.

When COAs are permuted to make multiple COAs, the permutation should be
done \emph{before} any dropping of items, not after.
}
