% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders1.R
\name{genAllOrders}
\alias{genAllOrders}
\title{Generate all possible orderings}
\usage{
genAllOrders(nvars)
}
\arguments{
\item{nvars}{Number of variables for which to generate orderings, which will
be the length of each ordering.}
}
\value{
Array of dimension \code{nvars} by \verb{nvars!}, with one ordering
per column.
}
\description{
This function generates all possible orderings of as many variables
as specified.  It is really just a wrapper for \code{\link[combinat:permn]{combinat::permn()}}, as
"orderings" are permutations.
}
\details{
The function warns if called with \code{nvars>10} but attempts it anyway.
Memory problems are likely in that case.

While useful in computing exact Shapley Values, using all orderings is
much less efficient than a combinations-based approach, and not
feasible in problems with more than about 12 items or variables.
}
