% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders1.R
\name{genCyclic}
\alias{genCyclic}
\title{Generate cyclic orderings}
\usage{
genCyclic(nvars, norders = 1, ncycles = NULL, random = FALSE)
}
\arguments{
\item{nvars}{Number of variables for which to generate orderings, which will
be the length of each ordering.}

\item{norders}{Number of orderings (minimum) to generate.}

\item{ncycles}{Number of full cycles to generate; overrides \code{norders} if
not \code{NULL}.}

\item{random}{If \code{TRUE}, first set of cycles should be random, not \code{1:n}.}
}
\value{
Array of dimension \code{nvars} by (at least) \code{norders},
with one ordering per column.
}
\description{
This function generates cyclic orderings, as many as desired, for as many
variables as desired.
}
\details{
This generates only full cycles, so the number of orderings may slightly
exceed the number requested by \code{norders} when \code{norders/nvars} is not an
integer.

When multiple cycles are requested, each is random, with no attempt to
prevent duplicates.  This is intended for use in research and play, not
recommended for "live" data use.
}
