% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders1.R
\name{genRandOrders}
\alias{genRandOrders}
\title{Generate random orderings}
\usage{
genRandOrders(nvars, norders)
}
\arguments{
\item{nvars}{Number of variables for which to generate orderings, which will
be the length of each ordering.}

\item{norders}{Number of orderings to generate.}
}
\value{
Array of dimension \code{nvars} by \code{norders}, with one ordering
per column.
}
\description{
This function generates random orderings, as many as desired, for as many
variables as desired.
}
\details{
This does not prevent generation of duplicates, which should be an
issue only for large numbers of orders on few variables.  It is intended
for use in researching COA performance vs. random, not recommended for
"live" data use.
}
