% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders1.R
\name{genWilliams}
\alias{genWilliams}
\title{Generate Williams designs of orderings}
\usage{
genWilliams(nvars, norders = NULL, ...)
}
\arguments{
\item{nvars}{Number of variables for which to generate orderings, which will
be the length of each ordering.}

\item{norders}{Number of orderings to generate.}

\item{...}{
  Arguments passed on to \code{\link[=genCyclic]{genCyclic}}
  \describe{
    \item{\code{ncycles}}{Number of full cycles to generate; overrides \code{norders} if
not \code{NULL}.}
    \item{\code{random}}{If \code{TRUE}, first set of cycles should be random, not \code{1:n}.}
  }}
}
\value{
Array of dimension \code{nvars} by (at least) \code{norders}, with one
ordering per column.
If \code{ncycles} is specified, there will be \code{nvars * 2 * ncycles} columns.
}
\description{
This function generates Williams orderings, as many as desired, for as
many variables as desired.
}
\details{
Williams designs are a cyclic set of orderings plus its "fold-over" or
reverse orderings.  Each design has 2 x nvars orderings.

The function generates only full cycles, so the number of orderings may
slightly exceed the number requested by norders when norders/nvars/2
is not an integer.

When  multiple cycles are requested, each is random, with no
attempt to prevent duplicates.  This is intended for use in research
and play, not recommended for "live" data use.
}
