% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{getCOA}
\alias{getCOA}
\title{Generate COA or COAs in final form for use elsewhere}
\usage{
getCOA(nvars, ncoa = 1, sep = FALSE, random = FALSE, nopermute = FALSE, ...)
}
\arguments{
\item{nvars}{The number of items/variables needed in the COA(s).  Cannot
exceed 1000.}

\item{ncoa}{Number of COAs to generate; see \code{sep} below.}

\item{sep}{Ignored when \code{ncoa=1}.  If \code{TRUE}, return a list of separate COAs.
If \code{FALSE}, return a single "stacked" COA combining the \code{ncoa}
separate ones.}

\item{random}{If \code{TRUE}, pick a random initial permutation to form the
COA, meaning the result will NOT reproduce on future runs.}

\item{nopermute}{If \code{TRUE}, the original algorithmic COA's columns are
\emph{not} permuted, resulting in a COA with very bad triplet imbalance.
This should never be used except for research or play.}

\item{...}{
  Arguments passed on to \code{\link[=getOA]{getOA}}
  \describe{
    \item{\code{nocheck}}{Whether to skip running \code{\link[=ckOA]{ckOA()}} on the generated OA.
Running the check can be time-consuming and is only needed in early
debugging.}
    \item{\code{silent}}{Logical, passed on to \code{\link[=ckOA]{ckOA()}} if \code{nocheck} is FALSE}
  }}
}
\value{
A COA in 1-based ncol>nrow format as expected by most analysis
drivers.  If \code{ncoa=1}, it will have \code{nvars} rows and at least
\code{nvars * (nvars-1)} columns, more if \code{nvars} is not a prime or power
of a prime.  If \code{ncoa>1}, the result will be correspondingly larger if
\code{sep} is \code{FALSE}, or a list of such COAs if \code{sep} is \code{TRUE}.
}
\description{
This is the function intended for end-users to generate a COA, or a group
of non-overlapping ones, or a stacked COA made from the non-overlapping
ones.
}
\details{
If \code{nvars} is not prime or a power of a prime, this function reduces the
base COA to the right size.
Note that the set of COAs generated by \code{ncoa>1} is \emph{never} reproducible,
unless the caller manages \code{set.seed} to make it so.  A single COA, however,
is automatically reproducible from call to call unless \code{random} is \code{TRUE}.
}
