% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{getOA}
\alias{getOA}
\title{Generate smallest orthogonal array for a given number of items/variables}
\usage{
getOA(nvars, nocheck = FALSE, silent = FALSE, random = FALSE)
}
\arguments{
\item{nvars}{The number of items/variables needed.}

\item{nocheck}{Whether to skip running \code{\link[=ckOA]{ckOA()}} on the generated OA.
Running the check can be time-consuming and is only needed in early
debugging.}

\item{silent}{Logical, passed on to \code{\link[=ckOA]{ckOA()}} if \code{nocheck} is FALSE}

\item{random}{Logical, item numbers in the OA are randomly relabeled if
\code{TRUE}.  This relabeling does \emph{not} make the resulting OA or any COAs
derived from it different in any practically useful way, and is \emph{not} a
substitute for \code{\link[=permuteCOA]{permuteCOA()}}.}
}
\value{
An orthogonal array, with items numbered 0 to \code{nvars-1},
\code{nvars} or slightly more (depending on next-largest prime or power)
columns and as many rows as the square of the number of columns.
This is the "0-based, nrow>ncol" format for an OA.
}
\description{
This function generates the smallest Galois orthogonal array large enough
to handle a given number of items or variables.
}
