% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{increments}
\alias{increments}
\title{Turn ordered result value measures into incremental values}
\usage{
increments(result)
}
\arguments{
\item{result}{A  vector, matrix or 3-D array (one result vector
per column) of results to be differenced.
Note that the result should be in order by steps taken,
not by original variable numbers.}
}
\value{
A vector or matrix of same shape and size as the parameter,
of increments across the positions.
}
\description{
This internal function "diffs" a set of total combo values to obtain
the increment added by each item.  It presumes a starting value of
zero, and tolerates even single-value results
(i.e., for a single combination size).
}
\keyword{internal}
