% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{KDRsolve1Combo}
\alias{KDRsolve1Combo}
\title{Solve a linear KDR for one combination of variables and one dependent
variable}
\usage{
KDRsolve1Combo(combo, scpMatrix, depvar)
}
\arguments{
\item{combo}{An vector of variable numbers (rows/columns of \code{scpMatrix})
specifying the variables in the combination to be evaluated.}

\item{scpMatrix}{A sum-of-cross-products matrix, from which either
a) the constant term in the regression has already been swept out
(easily accomplished by pre-centering all variables before creating
the SCP matrix) or b) the constant variable is included in \code{combo}.}

\item{depvar}{The row/column number in \code{scpMatrix} of the dependent variable.}
}
\value{
The (unadjusted) r-squared for the specified regression.
}
\description{
This function is a very simple (and not very efficient) illustration of
a valuation function.  It finds the linear regression r-squared for a
single dependent variable regressed on a single subset/combination of
variables.  It is intended for use with \code{\link[=SVsByCombos]{SVsByCombos()}}.
}
\details{
For practical use, see \code{\link[=KDRsolveCombos]{KDRsolveCombos()}}.  Inspect the source of this function
to see just how simple a valuation function can be.

It is more efficient to handle many combinations at once, so as to avoid
repeated function calls.  If there are multiple dependent variables to be
done, it is more efficient in a linear regression to handle them in
parallel as well.  But those efficiency issues are not large.  However,
putting the key code in C++, as \code{\link[=KDRsolveCombos]{KDRsolveCombos()}} effectively does,
gives a huge efficiency gain.
For larger problems, the valuation function needs to be very
efficient, especially when used with the \code{\link[=SVsByCombos]{SVsByCombos()}} driver that gives
exact Shapley Values.
}
