% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{KDRsolveComboR}
\alias{KDRsolveComboR}
\title{Solve slowly for linear r-squared for multiple combinations of variables}
\usage{
KDRsolveComboR(combo, scpMatrix, depvar)
}
\arguments{
\item{combo}{An matrix of variable numbers (rows/columns of \code{scpMatrix})
specifying the variables in the combinations to be evaluated,
one column per combination,
and with entries giving which rows/columns of the SCP matrix (next
parameter) the desired variables are in.}

\item{scpMatrix}{A sum-of-cross-products matrix, from which the constant
term in the regression has aleady been swept out.  (Easily accomplished by
pre-centering all variables before creating the SCP matrix.)}

\item{depvar}{A vector of the row/column numbers in \code{scpMatrix} of the
dependent variables.}
}
\value{
If there are multiple dependent variables (\code{length(depvar)>1}),
an array of r-squared values, one row per column in \code{combo} and one column
per dependent variable.  For a single dependent variable, a vector
of r-squared values, one per combination in \code{combo}.
}
\description{
This function finds r-squareds of regressions on multiple combinations
of one or more independent variables.  All combinations must be of the
same size.  It is intended for use with \code{\link[=SVsByCombos]{SVsByCombos()}}.
\code{\link[=KDRsolveCombos]{KDRsolveCombos()}} is faster, as it does the same things in C++.
}
