% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{KDRsolveCombos}
\alias{KDRsolveCombos}
\title{(Adjusted) R-squared for linear regression on a set of combinations}
\usage{
KDRsolveCombos(combo, scpMatrix, depvar, adjusted = NULL, both = FALSE)
}
\arguments{
\item{combo}{An integer matrix of combinations to solve for, one column
per combination and as many rows as the size of the combinations (all of
which must be the same size on any one call).}

\item{scpMatrix}{A sum-of-cross-products matrix, from which the constant
term in the regression has aleady been swept out.  (Easily accomplished by
pre-centering all variables before creating the SCP matrix.)}

\item{depvar}{For KDRsolveCombos_cpp, an integer vector of the row/column
positions in \code{scpMatrix} of the dependent variables
(numbered \verb{1..N} as in R).
For KDRsolveCombos1_cpp, a single integer indicating
the only dependent variable.}

\item{adjusted}{If not \code{NULL}, computes adjusted r-squared instead of
regular/straight r-squared.
Can be a numeric value giving the total degrees of freedom (DF)
to use for the adjustment, or can be \code{TRUE},
meaning to use the \code{ESS} (effective sample size) \emph{attribute} of
\code{scpMatrix}, created by \code{\link[=cpBuild]{cpBuild()}}.}

\item{both}{If \code{TRUE}, computes both straight and adjusted r-squareds.}
}
\value{
For a single dependent variable and \code{both=FALSE}, a vector of
r-squared values, one per column of \code{combo}.  Otherwise, a matrix with as
many rows as \code{combo} has columns, and a column for each dependent variable,
followed by, if \code{both} is \code{TRUE}, a column of adjusted r-squareds for
each dependent variable.
}
\description{
This is a valuation function for straight and/or adjusted r-squared for
linear regression on one or multiple dependent variables.

This function is designed to
be called from \code{\link[=SVsByCombos]{SVsByCombos()}}, and uses two
C++ functions to do the heavy lifting efficiently.
}
\details{
Multiple combinations, all of the same size, are handled in a single call,
so \code{multi=TRUE} should be specified to \code{\link[=SVsByCombos]{SVsByCombos()}} when it
is told to use this function.
}
