% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KDRsolveCombos_cpp}
\alias{KDRsolveCombos_cpp}
\alias{KDRsolveCombos1_cpp}
\title{Find r-squareds per combination for linear regression}
\usage{
KDRsolveCombos_cpp(combo, scpMatrix, depvar)

KDRsolveCombos1_cpp(combo, scpMatrix, depvar)
}
\arguments{
\item{combo}{An integer matrix of combinations to solve for, one column
per combination and as many rows as the size of the combinations (all of
which must be the same size on any one call).}

\item{scpMatrix}{A sum-of-cross-products matrix, from which the constant
term in the regression has aleady been swept out.  (Easily accomplished by
pre-centering all variables before creating the SCP matrix.)}

\item{depvar}{For KDRsolveCombos_cpp, an integer vector of the row/column
positions in \code{scpMatrix} of the dependent variables
(numbered \verb{1..N} as in R).
For KDRsolveCombos1_cpp, a single integer indicating
the only dependent variable.}
}
\value{
For \code{KDRsolveCombos_cpp}, a matrix of r-squareds found, with
one column per dependent variable and one row per combination in combo.

For \code{KDRsolveCombos1_cpp}, a vector of r-squareds found, one value for each
combination in combo.
}
\description{
\code{KDRsolveCombos_cpp} finds the linear regression r-squared for each of many
combinations of variables.  It can handle multiple dependent variables
at once.

\code{KDRsolveCombos1_cpp} does the same but is somewhat faster when there is only one
dependent variable.
}
\details{
\code{\link[=KDRsolveCombos]{KDRsolveCombos()}} is a friendlier wrapper for these functions; it
adds adjusted r-squareds if desired and automatically chooses which
of these to call.
}
