% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{KDRsolveOrder1}
\alias{KDRsolveOrder1}
\title{Single-order KDR solution, for testing purposes only!}
\usage{
KDRsolveOrder1(order, ...)
}
\arguments{
\item{order}{A vector of item orders, that must include all items
(i.e., it cannot be a size-limited ordering).}
}
\value{
A 3-D array of KDR values, dimensioned (nitems,1,nres)
where nres is the number of results (e.g., raw and adjusted r-squareds)
specified by the ... parameters to \code{KDRsolveOrders_cpp()}.
}
\description{
This function gets a set of KDR solutions for a single-ordering only,
by invoking the usual multi-ordering routine with a one-column matrix
of orderings.

It is used only to test the single-ordering capability of \code{SVsbyOrders()}
}
\keyword{internal}
