% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KDRsolveOrders_cpp}
\alias{KDRsolveOrders_cpp}
\title{Compute r-squareds for many different orders of variables}
\usage{
KDRsolveOrders_cpp(orders, scpMatrix, ndep = 0L)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column,
often from \code{\link[=getCOA]{getCOA()}}.}

\item{scpMatrix}{Cross-products matrix, without a constant
(either built with all pre-centered variables, or with the constant
already swept out and removed from the matrix).
The dependent variable(s) must be the last row/column(s) in \code{scpMatrix}.
See \code{\link[=cpBuild]{cpBuild()}}.}

\item{ndep}{Integer number of dependent variables.  If zero, it is
calculated as the number of columns in \code{scpMatrix} minus the
number of rows in \code{orders} (i.e., assuming all items in all orders).}
}
\value{
An array of R-squared values, with first two dimensions
matching those of the \code{orders} parameter and a third dimension
equal to \code{ndep}.
}
\description{
This function computes r-squared for each step, in order, in a list
of variables, for many different orderings of variables.  It can
handle multiple dependent variables.
}
\details{
This function combines an orderings-based driver with a regression-based
valuation function in C++ and
is the fastest option in the \code{ShapCOA} package to get
results for orderings-based linear key driver regressions.

It does not directly produce adjusted r-squareds or Shapley Values,
but the r-squareds it returns are used by its usual caller to get those.

It can handle size-limited Shapley Values in that the length of each
ordering need not include all variables (so, length is the
\code{maxitems} of some other functions) and the caller can enforce
the \code{minitems} idea.
}
