% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{makeCOA}
\alias{makeCOA}
\title{Generate a single "true" Component Orthogonal Array (COA)}
\usage{
makeCOA(oades, ...)
}
\arguments{
\item{oades}{Either the number of items/variables for which a COA is desired
OR an orthogonal array to be used, in the 0-based nrow>ncol format.}

\item{...}{
  Arguments passed on to \code{\link[=getOA]{getOA}}
  \describe{
    \item{\code{nocheck}}{Whether to skip running \code{\link[=ckOA]{ckOA()}} on the generated OA.
Running the check can be time-consuming and is only needed in early
debugging.}
    \item{\code{silent}}{Logical, passed on to \code{\link[=ckOA]{ckOA()}} if \code{nocheck} is FALSE}
    \item{\code{random}}{Logical, item numbers in the OA are randomly relabeled if
\code{TRUE}.  This relabeling does \emph{not} make the resulting OA or any COAs
derived from it different in any practically useful way, and is \emph{not} a
substitute for \code{\link[=permuteCOA]{permuteCOA()}}.}
  }}
}
\value{
A Component Order of Addition (COA) design, in the 0-based
nrow>ncol format, whose size matches that of the supplied OA in \code{oades} or
is at least as large as the number specified by \code{oades}.
}
\description{
This function generates a singleton pure COA, either from scratch or from
a supplied orthogonal array.
See Yang, Sun and Xu 2021 (Technometrics May 2021, 63:2).
}
\details{
The COA generated is exactly according to the Yang, Sun and Xu algorithm.
It may be larger than requested if the requested size is not a prime or
power of a prime, does \emph{not} have columns permuted to alleviate
triple imbalance and is not checked by \code{\link[=ckCOA]{ckCOA()}}.

This function is for internal use, only, by callers who handle the drops,
permutations and checking.
}
\keyword{internal}
