% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{permuteCOA}
\alias{permuteCOA}
\title{Permute columns of a COA to generate a new, non-overlapping, COA}
\usage{
permuteCOA(coades, ncoa = 2)
}
\arguments{
\item{coades}{The COA from which permutations are to be generated, in
0-coded nrow>ncol format.}

\item{ncoa}{The number of permutations to produce.  If less than 2,
the original \code{coades} is returned as the result.}
}
\value{
A list, each element of which is a COA, with the first element in
the list being the original \code{coades}.
The COAs are in the 0-based nrow>ncol format.
}
\description{
This function permutes a given COA to create a new one, none of whose
orderings are in the original.  It can produce as many permuted versions
as desired, with no overlaps among any of them.  Each permuted version has
the same equi-positioning and equi-precedence properties as the original.
}
\details{
The new permutatoins are randomly generated on each call and thus are \emph{not}
reproducible across runs unless the caller manages the random \code{set.seed}.

If a single new permutation is wanted, use \code{ncoa=2} and drop the first
COA returned (which is the original \code{coades}).

When COAs are permuted to make multiple COAs, the permutation should be
done \emph{before} any dropping of items, not after.  This function complains
if asked to permute a COA that appears to have had items dropped.
}
