% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{primeSize}
\alias{primeSize}
\title{Find smallest prime or power of a prime of at least a given size}
\usage{
primeSize(nvars, all = FALSE)
}
\arguments{
\item{nvars}{The number of items/variables needed.}

\item{all}{if \code{TRUE}, returns all primes/powers up to \code{nvars}.}
}
\value{
For \code{all} \code{FALSE}, an integer prime or power.  For \code{all} \code{TRUE},
an integer vector of primes and powers, all less than \code{nvars}.`
}
\description{
This function finds the smallest size of an orthogonal array (OA) or
Component Order of Addition (COA) design that is large enough
for a given number of items or variables.  A valid size must be a prime
number or a power of a prime number.
}
\details{
This function can be used to pre-determine the size of OA or COA that will
be needed for a given problem, or to generate a list of sizes for
research, play or presentation purposes.
}
