% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{SVsByCombos}
\alias{SVsByCombos}
\title{Exact, combinations-based, Shapley Values on an arbitrary valuation function}
\usage{
SVsByCombos(
  vfunc = KDRsolveCombos,
  nitems,
  minitems = 0,
  maxitems = nitems,
  multi = TRUE,
  nonull = TRUE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{vfunc}{The valuation function (see below).}

\item{nitems}{The number of items or variables in the problem.}

\item{minitems}{The minimum combination size to consider, defaulting to
zero.  If this is 2, for example, value increments from combinations of
size 2 to 3 are included in the calculation, but increments from size 0
to 1 and size 1 to 2 are not.}

\item{maxitems}{The maximum combination size to consider, defaulting to
all items.  If this is 5, for example, value increments from size 4
to 5 are included, but from size 5 to 6 are not.
If 0 or \code{NULL}, the value of \code{nitems} is used.}

\item{multi}{Whether \code{vfunc} can be called with a matrix of many combinations
of a given size, or if \code{FALSE}, must be called for each combination
individually.  Solutions are a bit faster with value functions written to
allow this.
If \code{vfunc} is \code{\link[=KDRsolveCombos]{KDRsolveCombos()}}, \code{multi} must be \code{TRUE}.}

\item{nonull}{Whether \code{vfunc} can/should be called for the \code{NULL} combinaton,
or a value of zero should simply be assumed in that case.
If \code{vfunc} is \code{\link[=KDRsolveCombos]{KDRsolveCombos()}}, \code{nonull} must be \code{TRUE}.}

\item{silent}{Whether to print progress/information messages to console
as the function works.}

\item{...}{Additional parameters to pass to the valuation function \code{vfunc}.}
}
\value{
A matrix of exact Shapley Values on the \code{vfunc} value increments,
one column per result type provided
by \code{vfunc}, and one row per item.  If only one result is provided by
\code{vfunc}, a vector with one value per item.
}
\description{
This function calculates exact Shapley Values using the combinations
approach, for any user-supplied valuation function.
}
\details{
This is a general driver for the combinations approach.  That approch is
typically not computationally feasible for more than 25 to 30 items, but
it does yield exact, not approximate, Shapley Values.

For linear Key Driver Regressions,
\code{\link[=KDRsolveCombos]{KDRsolveCombos()}} is a natural counterpart to this function.
\subsection{The Valuation Function:}{

The first argument to the value function is the combination(s) it must
compute for.  Further arguments are provided through the \code{...} mechanism.

If \code{multi} is 'FALSE`, the value function will be called once for each
combination, with its first parameter being the combination \emph{vector}.
It would usually return a single numeric result.

The value function can return multiple measures, if desired, as a vector.
The number of measures need not be specified in parameters;
it is determined by a preliminary dummy call to the function.

If \code{multi} is TRUE, the value function will be called with an entire
\emph{matrix} of combinations, one per column.  On any one call, all
combinations will be of the same size.
The value returned is usually a vector with one value per combination.
one row per position in the orderings and one column per ordering
In the case of multiple measures, the return value is an array with
one row per combination and one column per measure.

Some valuation functions cannot handle the \code{NULL} combination and many do
not need to.  But logistic regression, for example, might need to evaluate
a NULL combo to establish a non-zero log-likelihood starting point for
other variables to move from.  The \code{nonull} parameter specifies whether
to call the value function with a null combination, as opposed to just
assuming its value is zero.
}
}
