% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{SVsByOrders}
\alias{SVsByOrders}
\title{Approximate orderings-based Shapley Values on an arbitrary valuation
function using COAs or other orderings}
\usage{
SVsByOrders(
  orders,
  vfunc = KDRsolveOrders_cpp,
  silent = FALSE,
  minitems = 0,
  maxitems = 0,
  multi = FALSE,
  ...
)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column, often a COA
from \code{\link[=getCOA]{getCOA()}}.}

\item{vfunc}{The valuation function (see below).}

\item{silent}{If TRUE, does not comment on number of values returned
by the value function \code{vfunc}.}

\item{minitems}{Minimum combination size to consider, defaulting to zero.
If this is 2, for example, contributions from the first two items
in each ordering are excluded.}

\item{maxitems}{Maximum combination size to consider, treated as unlimited
if zero or \code{NULL}.}

\item{multi}{Whether \code{vfunc} can be called with a matrix of many orderings
at once, or if \code{FALSE}, must be called for each ordering
individually.  Solutions are (usually trivially) faster with
value functions written to allow this.}

\item{...}{Additional parameters to pass to the valuation function \code{vfunc}.}
}
\value{
A vector or matrix, depending on which \code{vfunc} returns, with the
approximate Shapley Values for each item/variable in the orderings and one
column, if a matrix, for each measure returned by \code{vfunc}.
}
\description{
This function calculates approximate Shapley Values using the orderings
approach, for an appropriate user-supplied valuation function.
}
\details{
The \code{orders} need not be full-length (i.e., need not have every item
occuring in every ordering), so this works for "size-limited Shapley
Values".  The number of items is deduced from the maximum value appearing
in \code{orders}.  If \code{maxitems} is specified but \code{orders} are full-length,
only the first \code{maxitems} rows are used.
\subsection{The Valuation Function:}{

The first argument to the value function is the ordering(s) it must
compute for.  Further arguments are provided through the \code{...} mechanism.

If \code{multi} is 'FALSE\verb{, the value function will be called once for each  ordering, with its first parameter being the ordering vector.   It would usually return a vector of results, for each combination size as it steps through the ordering.  It must *NOT* return increments as opposed to total values; the differencing is done by this function.   It must *NOT* return values in item-number order, but must keep the ordering order.  In other words, if item 5 is the first in an ordering, the first value returned by }vfunc` must be the contribution of item
5 as the first one in.

The value function can return multiple measures, if desired, as a matrix
with one column per measure (e.g., raw and adjusted r-squared, or
results for two different dependent variables).  The number of measures
need not be specified in parameters; it is determined by a preliminary
dummy call to the function.

If \code{multi} is TRUE, the value function will be called with the entire
orders matrix (as well as one test call with a single ordering as a
single-column matrix).  The value returned is usually a matrix with
one row per position in the orderings and one column per ordering
(in other words, the same shape as the \code{orders} matrix).  In the case
of multiple measures, the return value has a third dimension, which is
the number of measures.
}
}
