% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{SVsByOrders1}
\alias{SVsByOrders1}
\title{Simplest possible approximate orderings-based Shapley Values on an
arbitrary combination-based value function using COAs or other orderings}
\usage{
SVsByOrders1(orders, vfunc, minitems = 0, silent = FALSE, ...)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column, often a COA
from \code{\link[=getCOA]{getCOA()}}.}

\item{vfunc}{The valuation function (see below).}

\item{minitems}{Minimum combination size to consider, defaulting to zero.
If this is 2, for example, contributions from the first two items
in each ordering are excluded.}

\item{silent}{Whether to comment to console
when implied maximum combination size is not same as number of items.}

\item{...}{Additional parameters passed to the value function.}
}
\value{
A numeric vector of the average value increments, which is an
approximation of the Shapley Values if each ordering covers all
variables and the orderings are some type of random sample (e.g., a COA).
}
\description{
This function evaluates a user-supplied value function over multiple
orderings of items or variables (likely COAs) and averages the results to
obtain approximate Shapley Values.

This function will work for any value function that can evaluate
a single combination, the simplest possible requirement for
a value function.
But, it is \strong{not efficient}  in that such a value function can
never make any use of any efficiencies
in successive "steps into" an ordering.
}
\details{
The \code{orders} need not be full-length (i.e., need not have every item
occuring in every ordering), so this works for "size-limited Shapley
Values".  The number of items is deduced from the maximum value appearing
in \code{orders}.  Unlike some other drivers, this one does \emph{not} shorten
orders according to a \code{maxitems} parameter.

Minimum combination sizes are supported by the \code{minitems} parameter.
\subsection{The Valuation Function:}{

The first argument to the value function is a vector giving the combination
it must compute for.
Further arguments to it are provided through the \code{...} mechanism.

Multiple measures per combination are \strong{not} supported by this driver;
the value function must return a single numeric value.
}
}
