% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{SVsForKDRsByOrders}
\alias{SVsForKDRsByOrders}
\title{Approximate Shapley Values in Key Driver Regressions using COAs
or other orderings}
\usage{
SVsForKDRsByOrders(
  orders,
  scpMatrix,
  ndep = 0,
  adjusted = NULL,
  both = FALSE,
  minitems = 0,
  maxitems = 0
)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column, often a COA
from \code{\link[=getCOA]{getCOA()}}.}

\item{scpMatrix}{Cross-products matrix, \emph{without} a constant
(either built with all pre-centered variables, or with the constant
already swept out and removed from the matrix).
The dependent variable(s) must be the last row/column(s) in \code{scpMatrix}.
See \code{\link[=cpBuild]{cpBuild()}}.
If \code{adjusted == TRUE}, it must contain
an \code{ESS} attribute specifying the total number of respondents or
effective sample size used to create it.}

\item{ndep}{Integer number of dependent variables.  If zero, it is
calculated as the number of columns in \code{scpMatrix} minus the
number of rows in \code{orders} (i.e., assuming all items in all orders).
This won't work when \code{maxitems} is less than the number of rows
in \code{orders}.}

\item{adjusted}{If numeric and not zero, requests the adjusted r-squared
and gives the total d.f. to use for it.
If TRUE, requests adjusted r-squared using d.f. from the \code{ESS} attribute
of the \code{scpMatrix} argument.}

\item{both}{If \code{TRUE}, compute raw r-squared in addition to adjusted.}

\item{minitems}{Minimum combination size to consider, defaulting to zero.
If this is 2, for example, contributions from the first two items
in each ordering are excluded.}

\item{maxitems}{Maximum combination size to consider, treated as unlimited
if zero or \code{NULL}.}
}
\value{
A vector or matrix, depending on which \code{vfunc} returns, with the
approximate Shapley Values for each item/variable in the orderings and one
column, if a matrix, for each measure returned by \code{vfunc}.

A vector (typically) or matrix (if \code{both}, or if \code{ndep>1})
of R-squared (or adjusted) increments.
If a matrix, columns correspond to the dependent variables, and
if \code{both}, further columns for adjusted r-squareds, in the same order.
}
\description{
This function computes SVs for r-squared and/or adjusted r-squared in
linear key driver regressions.
It can compute for multiple dependent variables
and can handle size-limited Shapley Values.
}
\details{
This is a wrapper around \code{\link[=KDRsolveOrders_cpp]{KDRsolveOrders_cpp()}} that
adds the options of adjusted r-squared and size-limited SVs to
the results.
It is the fastest option in the \code{ShapCOA} package to get
COA-based approximate Shapley Values on KDRs.

The \code{orders} need not be full-length (i.e., need not have every item
occuring in every ordering), so this works for "size-limited Shapley
Values".  The number of items is deduced from the maximum value appearing
in \code{orders}.  If \code{maxitems} is specified but \code{orders} are full-length,
only the first \code{maxitems} rows are used.

In the (unlikely) event that \code{orders} gives all possible orderings,
the SVs computed are exact.
}
