% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVsForMaxDiff.R
\name{SVsForMaxDiff}
\alias{SVsForMaxDiff}
\title{Approximate Shapley Values for TURF Analysis on MaxDiff data}
\usage{
SVsForMaxDiff(
  orders = NULL,
  utilities,
  xform = NULL,
  anchor = FALSE,
  tasksize = 0,
  threshold = TRUE,
  weights = NULL,
  minitems = 0,
  maxitems = 0,
  ncoas = 1,
  ...
)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column, often a COA
from \code{\link[=getCOA]{getCOA()}}.
If \code{NULL}, a COA will be generated via \code{\link[=getCOA]{getCOA()}}, using the
\code{ncoa} parameter.`}

\item{utilities}{MaxDiff logit \emph{utilities} (\strong{not} transformed \emph{scores}).
Need not be zero-centered or zero-anchored.  Alternatively, for
standard 0-1 TURF, the \code{0/1} scores.}

\item{xform}{How to transform MaxDiff utilities into scores for analysis.
See details below.  In brief, 0 means MNL,
1 means default individual utility probability score,
2 means anchored probability score, and
NULL means use sums of unexponentiated utilities (i.e., 0/1 data).}

\item{anchor}{If numeric, the column number of the anchor item
in an anchored MaxDiff;
it will be used for centering and then removed.
If \code{FALSE}, \code{utilities} are not from anchored data.
If \code{TRUE}, \code{utilities} are anchored but have \emph{already} been zero-anchored
and the anchor column has already been removed.}

\item{tasksize}{The number of items per task in the original MaxDiff tasks.
This is the "a" in the published Sawtooth formulas;
it is used in computing the weighted probability scores when \code{xform}
is 1 (unanchored MaxDiff).
Required if \code{xform} is \code{1}, ignored otherwise.}

\item{threshold}{If TRUE or numeric, and \code{xform} is not \code{NULL},
weighted probability scores are subjected to a \emph{post}-threshold,
either of the specified non-zero numeric value,
or of a standard default of \code{0.5} for MNL (\code{xform=0}),
\code{0.9} for unanchored WPS (\code{xform=1})
or \code{0.5} for anchored Maxdiff (\code{xform=2}).
If \code{FALSE}, TURF calculations are on the weighted probability scores
themselves, without thresholding.
If \code{xform} is NULL, this is the "depth" parameter for a standard
\code{0/1} TURF.}

\item{weights}{Vector of respondent weights, one per row of \code{df}.
Unweighted data if \code{NULL}.  Weights may be zero to subset data but should
never be negative.}

\item{minitems}{Minimum combination size to consider, defaulting to zero.
If this is 2, for example, contributions from the first two items
in each ordering are excluded.}

\item{maxitems}{Maximum combination size to consider, treated as unlimited
if zero or \code{NULL}.}

\item{ncoas}{How many COAs to "stack" for extra precision,
applicable only if \code{order} is \code{NULL} and the COA is generated by
this function.}

\item{...}{Available to pass \code{random} or \code{silent} parameters
onward to \code{\link[=getCOA]{getCOA()}}.  Irrelevant \code{orders} is supplied.}
}
\value{
A vector of approximate Shapley Values, one value per column
in \code{utilities}.
}
\description{
This function computes Shapley Values using COAs or other orderings of
items, for various forms of TURF on MaxDiff data
(and for standard 0-1 TURFs).
It uses the valuation function \code{\link[=SVsMaxDiffCalc_cpp]{SVsMaxDiffCalc_cpp()}}.
}
\details{
For standard 0-1 TURFs and many variants of it, including depth>1 and many
measures other than standard reach, this function is slow and unnecessary.

Three major versions of scoring combinations of MaxDiff items are supported,
specified by the \code{xform} parameter:
multinomial logit (MNL), the default Sawtooth individual probability score
and the (default for anchored MaxDiffs)
Sawtooth anchored probability score.

Two ways to measure the TURF "reach" are directly supported: the total
weighted probability scores (WPS), and whether the WPS clear a specified
threshold ("post-thresholding".

A third measurement option is standard 0-1 TURF resulting
from "pre-thresholding," also known as "item thresholding",
where a threshold is applied in advance to each individual item,
based on either utilities or some kind of score.
If this is done here, the \code{utilities} parameter must contain only \verb{0's}
and \verb{1's}. In other words, the pre-thresholding must be done by the caller;
it is not handled by this function.

Weighted data is supported, as are size-limited Shapley Values.

The \code{orders} need not be full-length (i.e., need not have every item
occuring in every ordering), so this works for "size-limited Shapley
Values".  The number of items is deduced from the maximum value appearing
in \code{orders}.  If \code{maxitems} is specified but \code{orders} are full-length,
only the first \code{maxitems} rows are used.

\code{tasksize} only applies if \code{xform} is 1, since the default weighted
probability calculation uses the task size.
It is required in that case.
\subsection{MaxDiff Scores}{

\code{xform} controls how utilities are centered or anchored and how
they are transformed into scores for the TURF analysis:
\itemize{
\item {\code{xform=0   }}  MNL-based scoring.  Utilities are not centered.
After exponentiating, individual scores are re-percentaged to
add to \code{100}.  The score for a combination of items is the sum
of the individual item scores.
\item {\code{xform=1   }}  The default individual utility probability
scoring.  Utilities are first zero-centered.  The scores are
\code{E / ( E + a - 1 )}, where \code{E} is the exponentiated utility for
an item and \code{a} is the number of items
in the original tasks, given by \code{tasksize}.
For the score for a combination of items, \code{S} is the sum of the
exponentiated utilities for the items in the combination and
then calculated as \code{S / ( S + a - 1 )}.
\item {\code{xform=2   }}  Anchored probability scores.  The utilities
are assumed to have already been zero-anchored and are not
further centered.  Scores for items and combinations are
the same as for \code{xform=1} except that the final transformation
is \code{S / ( S + 1)} and the task size is not used.
}

Note that the scores used for the TURF are \emph{not} the same as the
"rescaled weighted probability scores" Sawtooth Software
uses for \emph{aggregate} reporting.
The default rescaled WPS uses the \code{xform=1} calculation for the items
detailed above, then rescales the results to add to 100 before averaging
over respondents.  The anchored WPS uses the same calculation
at the respondent level as the \code{xform=1} case
(so the tasksize does matter now) and then multiplies the result by
\code{100 / (1/a)} so that items with utility equal to the anchor get a final
score of \code{100} and the maxium score is \verb{100a}.
}
}
