% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVsMaxDiffCalc_cpp}
\alias{SVsMaxDiffCalc_cpp}
\title{Valuation function for various forms of TURF and TURF on MaxDiff analyses}
\usage{
SVsMaxDiffCalc_cpp(
  orders,
  utilities,
  weights,
  maxdiff = TRUE,
  xform = 1L,
  threshold = 0,
  tasksize = 0L
)
}
\arguments{
\item{orders}{An integer matrix of orderings, one per column,
typically from \code{\link[=getCOA]{getCOA()}}.}

\item{utilities}{The exponentiated MaxDiff utilities, in a numeric
matrix with one column per item and one row per respondent.
For standard \code{0/1} TURF, the \code{0} and \code{1}.}

\item{weights}{Vector of respondent weights, one per row of \code{utilities}.
Weights may be zero to subset data but should never be negative.
The parameter is required; use a vector of 1's for unweighted data.}

\item{maxdiff}{Logical, whether \code{utilities} contains exponentiated MaxDiff
utilities (the default).
If not, \code{0-1} reach data is assumed but is not checked.}

\item{xform}{How to transform total exponentiated utilities
scores for a combination into the final score for the combination.
Ignored if \code{maxdiff} is \code{FALSE}.
See \code{\link[=SVsForMaxDiff]{SVsForMaxDiff()}} for details.}

\item{threshold}{For MaxDiff, the numeric threshold to use for
post-thresholding of MaxDiff scores.
If \code{maxdiff} is \code{FALSE}, the "depth" to use in evaluating a standard
\code{0/1} TURF on "reach".}

\item{tasksize}{The number of items per task in the original MaxDiff tasks.
This is the "a" in the published Sawtooth formulas;
it is used in computing the weighted probability scores when \code{xform}
is 1 (unanchored MaxDiff).
Required if \code{xform} is \code{1}, ignored otherwise.}
}
\value{
A \code{double} matrix of combination values, with the same shape and
size as the \code{orders} matrix.
}
\description{
This function computes the the total value of combinations of various
forms of TURF analysis for multiple orderings of items.
It is most useful for TURF on MaxDiff data.
It is tightly coupled with \code{\link[=SVsForMaxDiff]{SVsForMaxDiff()}}, which handles the final
aggregation into Shapley Values, including differencing and
unrotating the orderings.
It relies on its caller to check parameters and center or exponentiate
utilities as needed.
Even though it is exported and available,
it is not \emph{intended} for use in other situations.
}
