% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{swp}
\alias{swp}
\title{Do a Gauss-Jordan "sweep" (aka Beaton's \code{swp} operator) on a single
row/column of a cross-products matrix}
\usage{
swp(CP, k)
}
\arguments{
\item{CP}{The cross-products matrix to be swept.  It may well have been
partially swept before.}

\item{k}{The number of the row/column to be swept.}
}
\value{
The swept cross-products matrix, with same dimensions as \code{CP}.
}
\description{
This function performs a single sweep or "swp" operation on a matrix.
}
\details{
This is a building block for full regressions, and for stepping through
variables one at a time.  It is more for testing and play than for any
real production use, as repeated calls over many orderings are very
inefficient and various integrated C++ functions handle this better.
}
\keyword{internal}
