% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drivers1.R
\name{totalup}
\alias{totalup}
\title{Total SV item contributions by variable, undoing the ordering they were in}
\usage{
totalup(pieces, orders)
}
\arguments{
\item{pieces}{Numeric array of results, one ordering per column,
one row per step in \code{orders} and a third dimension for multiple
measures obtained from the orderings.}

\item{orders}{The integer matrix of orderings the \code{pieces} follow,
with the same shape and size as the first two dimensions  of \code{pieces}.}
}
\value{
A vector or matrix (one fewer dimension than \code{pieces}, of totals
for each variable or item (as many as the maximum value in orders)
and as many columns as the third dimension of \code{pieces}.
}
\description{
This internal function totals contributions by each variable across all
orderings, effectively "unscrambling" the ordering of resultsof results.
}
\keyword{internal}
