% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turf1.R
\name{turf1}
\alias{turf1}
\title{Evaluate a standard 0-1 TURF on a single combination}
\usage{
turf1(comb, tdata, depth = 1, wts = 1)
}
\arguments{
\item{comb}{The combination to evaluate, a vector of item numbers.}

\item{tdata}{The 0/1 data, a matrix with one column per item
and one row per respondent.}

\item{depth}{The depth required for an overall reach
(i.e., the number of indiviudal item reaches required for an
overall reach).}

\item{wts}{The respondent weights, a vector with one element per row of
\code{tdata}, or a constant 1 for unweighted data.}
}
\description{
This is the simplest-minded possible TURF valuation function, intended
mostly for testing of various drivers.
}
\details{
Inputs are not checked in any way; the caller is just trusted.

Drivers should be told \code{multi=FALSE} and \code{nonull=TRUE} when applicable.
}
