% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COAs7.R
\name{userCode}
\alias{userCode}
\title{Convert orthogonal array or COA design to 1-based ncol>nrow format}
\usage{
userCode(des)
}
\arguments{
\item{des}{A design or orthogonal array (numeric matrix) to be converted.}
}
\value{
The design with 0-based coding,
and with runs or orders in the rows, and items in the
columns (nrow>ncol format).
}
\description{
This converts an orderings or combinations design to a format equivalent
to what standard R function \code{combn} produces, and the format the functions
using COAs expect.  That format:
\itemize{
\item uses 1-based numbering of items,
\item with runs/orders/combinations in the columns and items in the rows
(meaning the number of columns almost always exceeds the number of rows).
}

The design is examined to see which of the output criteria is does not
already meet.  An already-compliant design is returned unchanged.
}
\details{
Aside from development and testing, it is not clear what the use case for
this would be!
}
\seealso{
\code{\link[=designCode]{designCode()}} to do the reverse.
}
