% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{adjustOrNot}
\alias{adjustOrNot}
\title{Reconcile KDR adjusted and both flags, set totdf}
\usage{
adjustOrNot(adjusted = FALSE, both = FALSE, scpMatrix = NULL, maxvars = 0)
}
\arguments{
\item{adjusted}{If numeric and not zero, requests the adjusted r-squared
and gives the total d.f. to use for it.
If TRUE, requests adjusted r-squared using d.f. from the \code{ESS} attribute
of the \code{scpMatrix} argument.}

\item{both}{If \code{TRUE}, compute raw r-squared in addition to adjusted.}

\item{scpMatrix}{The SCP matrix to be used in the KDRs.
If \code{adjusted == TRUE}, it must contain
an \code{ESS} attribute specifying the total number of respondents or
effective sample size used to create it.}

\item{maxvars}{Total number of independent variables/items
(which the total d.f. must exceed).}
}
\value{
A list with three elements: doadj (logical, whether to compute
adjusted r-squared), doraw (logical, whether to do straight r-squared),
and totdf (total d.f. to use for adjusted r-squared, with 1 already
subtracted for the constant term).
}
\description{
Internal function for setting up the KDR adjusted and both flags, and
total degrees of freedom for adjusted r-squared.
}
\keyword{internal}
