% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDRs1.R
\name{checkWeights}
\alias{checkWeights}
\title{Check whether respondent weights are OK}
\usage{
checkWeights(weights, nresp, caller)
}
\arguments{
\item{weights}{The weights vector to be checked.}

\item{nresp}{Number of respondents and expected number of weights.}

\item{caller}{Character name of calling routine for use in any message
to console.}
}
\value{
\code{TRUE} if weights look OK, \code{FALSE} if any problem found.
}
\description{
This function prints a complaint message if weights are not of right
length or any are negative.
}
\keyword{internal}
